phase2: fix Devin Review findings on ea7516f (consume_if_once TOCTOU, upscale f64 precision, plugin scan mtime cache)
Real concurrency hardening:
- McpPermissionStore::consume_if_once: replace check (read lock) +
  grant (write lock) with a single write-lock-held read-modify-write.
  Two concurrent callers can no longer both observe the same Once
  grant and succeed. The MCP server's tiny_http accept loop is
  single-threaded so the race wasn't reachable in practice, but
  defence-in-depth at a permission gate is the right architecture.

FFI precision fix:
- upscale_lanczos: change scale parameter from f32 to f64 so values
  arriving from JavaScript (which only has f64 numbers) survive the
  FFI boundary intact. The previous `scale as f32` cast at the
  bridge layer silently rounded 1.0000001_f64 down to exactly 1.0_f32,
  which the algorithm then rejected as out of range. Internal kernel
  weights stay f32 (no need for 53-bit precision on pixel weights);
  only the scale parameter and resampling-center computation moved
  to f64.

Performance:
- PluginRegistry::scan: skip the full directory walk + manifest
  re-parse when nothing on disk has changed since the last scan.
  Mtime tracked on plugin_dir (for top-level add/remove) plus every
  manifest.json (for in-place edits). force_rescan() bypasses the
  cache when callers know more than the filesystem can tell us.

Per Devin Review:
- ANALYSIS_pr-review-job-0594c03f68c24589ba78a32926e3874f_0001 (consume_if_once)
- ANALYSIS_pr-review-job-0594c03f68c24589ba78a32926e3874f_0004 (upscale_lanczos)
- ANALYSIS_pr-review-job-0594c03f68c24589ba78a32926e3874f_0006 (plugin scan)

ANALYSIS_..._0002 (batch_start worker design) acknowledged as
correct — no code change.

cargo clippy --workspace --all-targets -- -D warnings : clean
cargo test --workspace --all-targets : all pass
pnpm typecheck && pnpm lint : clean

+++ crates/kcreate_bridge/src/phase2.rs
@@ -340,7 +340,7 @@ fn ai_models_dir() -> PathBuf {
     PathBuf::from(home).join(".kcreate").join("models")
 }
 
-pub fn ai_upscale(node_id: Uuid, scale: f32) -> Result<Uuid> {
+pub fn ai_upscale(node_id: Uuid, scale: f64) -> Result<Uuid> {
     // Load encoded image and decode.
     let (encoded, parent) = with_workspace(|ws| {
         let node = ws
